selftests: forwarding.config.sample: Show variable definitions as comments
The following patch will change lib.sh to source forwarding.config.sample
if forwarding.config is not available. So far it has been possible to
specify the configuration on the command line and through environment
variables for ad-hoc configuration. This would not work anymore, because
the config sample just sets the variables unconditionally.

But the variables actually do not need to be defined by forwarding.config,
because lib.sh contains a set of reasonable defaults. So instead, just show
the assignment in a comment. A user wishing to override the value would
uncomment the line and set whatever value they like.

Add empty lines between individual variables, because otherwise it's just
one huge block of comments, and is not clear anymore.

Signed-off-by: Petr Machata <petrm@nvidia.com>