selftests: forwarding: lib: Permit fallback to forwarding.config.sample
The forwarding selftests use a configuration file where names of interfaces
are defined and various variables can be overridden. There is also
forwarding.config.sample that users can use as a template to refer to when
creating the config file. What happens a fair bit is that users either do
not know about this at all, or simply forget, and are confused by cryptic
failures about interfaces that cannot be created.

Therefore emit a warning when forwarding.config is not available.

If no particular HW is tested, the sample configuration is usually just
fine. So when the sample configuration is found, use that instead.

It is also possible to specify interface names on the command line. In that
case, emit no warnings and trust the user that they have specified
everything to their liking through command line and environment variables.

Suggested-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Petr Machata <petrm@nvidia.com>