selftests: forwarding: Have RET track kselftest framework constants
The variable RET keeps track of whether the test under execution has so far
failed or not. Currently it works in binary fashion: zero means everything
is fine, non-zero means something failed. log_test() then uses the value to
given a human-readable message.

In order to allow log_test() to report skips and xfails, the semantics of
RET need to be more fine-grained. Therefore have RET value be one of
kselftest framework constants: 1 for fail, 2 for xfail, etc.

The current logic in check_err() is such that first non-zero value of RET
trumps all those that follow. But that is not right when RET has more
fine-grained value semantics. If an xfail is followed by a fail, the
resulting value should be fail.

Thus rewrite check_err() with that logic in mind. This will make it easier
to change check_err() to set xfail or fail depending on circumstances in a
later patch.

The function check_fail() needs a similar treatment. Instead, simply make
it a wrapper around check_err() with reverse error logic.

Signed-off-by: Petr Machata <petrm@nvidia.com>