selftests: forwarding: Convert log_test() to recognize RET values
In a previous patch, the interpretation of RET value was changed to mean
the kselftest framework constant with the test outcome: 1 for fail, 2 for
xfail, etc.

In this patch, update log_test() to recognize the various possible RET
values.

The RET outcome of the test influences EXIT_STATUS that the overall script
should return. Different outcomes have different weights thought: a skip
should only be reported if all tests are actually skipped. A success only
if none fails, etc. The ordering of the states as their "weight" increases
is skip, pass, xfail, fail. Add logic to merge the current RET into the
EXIT_STATUS.

For this to work, the initial value of EXIT_STATUS status needs to be
ksft_skip, so change it. This changes the exit code of a script that ran no
actual tests from 0 to ksft_skip. This should only be observable when
TESTS is overridden explicitly in the environment to be an empty string.
In that situation, exiting with ksft_skip seems apt.

With this log_test_skip can be reexpressed as a thin wrapper around
log_test.

Signed-off-by: Petr Machata <petrm@nvidia.com>