selftests: forwarding: Support for performance sensitive tests
Several tests in the suite use large amounts of traffic to e.g. cause
congestion and evaluate RED or shaper performance. These tests will not run
well on a slow machine, be it one with heavy debug kernel, or a VM, or e.g.
a single-board computer. Allow users to specify an environment variable,
KSFT_MACHINE_SLOW=yes, to indicate that the tests are being run on one such
machine.

Performance sensitive tests can then define PERF_SENSITIVE_TEST=yes so that
failures cause a ksft_xfail result, not a ksft_fail one. The value can be
specified on a check-by-check basis as well, for which there is a helper:

	perf_sensitive check_err $? "Expected much, got little."

Signed-off-by: Petr Machata <petrm@nvidia.com>