net: fib_rules: Add DSCP support
The 'tos' selector in ip-rule is implemented differently between IPv4
and IPv6. In IPv4 it is used to match on the three "Type of Services"
bits specified in RFC 791, while in IPv6 it is used to match on the six
DSCP bits (RFC 2474).

Add a new FIB rule attribute to allow matching on DSCP. The attribute
will be used to implement a 'dscp' selector in ip-rule with a consistent
behavior between IPv4 and IPv6.

Implement basic DSCP support in the common FIB rules code. Subsequent
patches will implement matching support in IPv4 and IPv6 and enable the
use of the attribute by adding it to the netlink policy.

Add an internal mask field and initialize it to cover the DSCP field in
the DS field when the new attribute is specified. The mask is needed to
differentiate between rules that match on a DSCP of zero and those that
do not match on DSCP at all.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>