bpf: Populate DSCP field in bpf_fib_lookup() helper
The helper performs a FIB lookup according to the parameters in the
'params' argument, one of which is 'tos'. According to the test in
test_tc_neigh_fib.c, it seems that BPF programs are expected to
initialize the 'tos' field to the full 8 bit DS field from the IPv4
header.

Therefore, extend the helper to populate the new DSCP field in the IPv4
flow info structure using the inet_dsfield_to_dscp() helper.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>