xfrm: Populate DSCP field in decode_session4()
The function populates the 'flowi4_tos' field in the IPv4 flow
information from the 'ip.tos' field in the flow keys structure. The flow
dissector populates this field with the entire 8 bits of the DS field.

Therefore, use inet_dsfield_to_dscp() helper to initialize the new DSCP
field in the IPv4 flow information structure.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>