ipv4: udp_tunnel: Populate DSCP in udp_tunnel_dst_lookup()
The function is called by drivers of various UDP tunnels (e.g., vxlan,
geneve) to resolve the route for the remote destination based on various
arguments. One of the arguments is 'tos' and all the callers treat it as
the DS field.

Therefore, use this value and the inet_dsfield_to_dscp() helper to
initialize the new DSCP field in the IPv4 flow information structure.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>