ipv4: icmp: Populate DSCP in icmp_route_lookup()
The function is used to resolve the route for a locally generated ICMP
packet according to provided arguments. One of the arguments is TOS
which will be set as the DS field in the encapsulating IPv4 header.

Therefore, use this value and the inet_dsfield_to_dscp() helper to
initialize the new DSCP field in the IPv4 flow information structure.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>