ipv4: Mask TOS in __fib_validate_source()
The function is called from the input path to make sure packets arrived
with a valid source and from the expected interface. The TOS value is
always provided to the input path with all but the three lower DSCP bits
masked. This is going to change in subsequent patches to allow matching
on the entire DSCP value. As a preparation for this change, mask all but
the three lower DSCP bits.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>