ipv4: icmp: Pass full DS field to ip_route_input()
All the callers of ip_route_input() pass the full DS field as the TOS
argument except the ICMP code. Align the ICMP code to the other callers
and pass the entire DS field. This should not result in a behavior
change since all but the three lower DSCP bits are masked by the input
path anyway.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>