ipv4: udp: Only mask ECN bits when calling ip_mc_validate_source()
The TOS argument is passed to ip_mc_validate_source() so that the source
of the multicast packet will be validated using fib_validate_source(),
but this function already masks all but the three lower DSCP bits.
Therefore, we can only mask ECN bits when calling
ip_mc_validate_source() with changing the behavior.

This will later allow us to extend fib_validate_source() to initialize
the new DSCP field in the IPv4 flow information structure.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>