ipv4: Only mask ECN bits when calling fib_validate_source()
The function already masks all but the three lower DSCP bits so there is
no need to perform this masking when calling the function. Only mask the
two ECN bits as this will later allow us to extend fib_validate_source()
to initialize the new DSCP field in the IPv4 flow information structure.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>