ipv4: Only mask ECN bits when calling ip_route_input_rcu()
All the users of the TOS argument already mask all but the three lower
DSCP bits so there is no need to perform this masking when calling the
function. Only mask the two ECN bits as this will later allow us to
extend the input path to initialize the new DSCP field in the IPv4 flow
information structure.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>