mlxsw: pci: Initialize XDP Rx queue info per RDQ
'struct xdp_rxq_info' was added to provide more info for XDP program.
The idea is to initialize the structure as part of init and attach it to
XDP buffer, so instead of updating N-members in xdp_buff, we can use the
same pre-initialized info. Rx info registration requires:
* netdevice
* Queue index
* NAPI ID

To pre-initialize such structure with all the info we need
{netdev} x {queue} pairs, as there is no queue for each netdevice using
Spectrum ASICs. This solution is optional, but creates huge overhead.

One option is to setup such structure per port, which means that queue
index and NAPI ID are not set. It's ok to not set them, but with such
option, the driver will not be able to set memory model later. Memory model
is necessary to allow recycling pages to the correct pool after redirect
action.

Another option is to setup such structure per Rx queue, and use dummy
netdevice, 'xdp_rxq_info->dev' is used to set 'xdp_md->ingress_ifindex',
which means that using dummy netdev, XDP program will not be aware of the
Rx port. The netdevice must be set, so dummy netdevice is a solution for
mlxsw driver.

An optional workaround can be to use Rx queue info per port, and set
rxq_info.mem.type directly, but then allocator is not initialized. This
solution works, but might create issues in the future.

Set XDP Rx queue info per RDQ, so then queue index and NAPI ID are
available, a future patch will set memory model also.

In future we can create many pairs of {netdev} x {queue}, so all the fields
will be set correctly.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>