mlxsw: Dedicate SDQ for XDP packets
A next patch will add support for transmitting XDP frames. There is a
limitation to call XDP APIs when budget is zero. In Rx flow, we just skip
processing for all packets. For Tx completion, driver should free the SKBs,
but can not handle XDP completions.

To be able to separate such handling, dedicate SDQ #1 for XDP frames. The
driver already dedicates SDQ #0 for EMADs. The appropriate CQ is also
dedicated for XDP, as it is dedicated for specific SDQ.

Add an enum for SDQ reserved indexes, instead of MLXSW_PCI_SDQ_EMAD_INDEX
macro. Derive minimum required SDQs from the new enum. Adjust the exiting
code to use the enum and to take into account the reserved SDQ for XDP.
Next patches will use this queue.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>