ipv6: fib_rules: Add DSCP mask matching
Extend IPv6 FIB rules to match on DSCP using a mask. Unlike IPv4, there
is no difference in matching between 'tos' and 'dscp' selectors, so the
DSCP mask is always initialized. Like IPv4, the mask is initialized to
cover the entire DSCP field if the mask attribute is not specified.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>