net: vxlan: Disallow changing multicast remote configuration when down
When a vxlan netdevice is brought up, if its default remote is a multicast
address, the device joins the indicated group.

Therefore when the multicast remote address changes, the device should
leave the current group and subscribe to the new one. Similarly when the
interface used for endpoint communication is changed in a situation when
multicast remote is configured. None of this is currently done.

Both vxlan_igmp_join() and vxlan_igmp_leave() can however fail. So should
we implement this fix, depending on whether old is left first or new is
joined first, we could end with an inconsistent configuration where neither
or both groups are joined.

Instead of dealing with these edge cases, when the remote is a multicast
address, simply forbid changing its address or the interface used for
endpoint communication. This changes behavior, however in a situation where
the current behavior is already broken.

Cc: Roopa Prabhu <roopa@nvidia.com>
Cc: Nikolay Aleksandrov <razor@blackwall.org>
Signed-off-by: Petr Machata <petrm@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>