selftests: mlxsw: vxlan_fdb_veto: Down VXLAN prior to changelink test
The changelink test verifies a mlxsw bounce. The following patch introduces
a bounce in the generic code that would mask the one that is tested for
here. Set the port down for the duration of the test -- that way the mlxsw
bounce stays visible even with the changes introduced in the next patch.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>