seg6: Extend seg6_lookup_nexthop() with an oif argument
seg6_lookup_nexthop() is a wrapper around seg6_lookup_any_nexthop()
which the previous patch extended with an oif argument. Extend
seg6_lookup_nexthop() with an oif argument and pass 0 from all the call
sites. This a preparation for the next patch that will allow user space
to specify an output interface for the End.X behavior.

No functional changes intended.

Reviewed-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>